src/data_loading/requirements.txt
pandas==0.24.2
google-cloud-storage==1.20.0